Remove redundant else statement
The previous code checked for if the specified .gol file was present,
and if not, created it and returned 0. The "else" statement following
was redundant because the preceding "if" portion would always return if
the condition was met.

Removing the "else" statement improves readability and reduces indents.
The control flow of the program is functionally equivalent after
removing the "else" statement.